pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "htb"
version = "0.0.1"
description = "H Trading Bot"
authors = [
    { name = "Matteo Johnston", email = "equilibria1610@gmail.com" },
]
readme = "README.md"
dependencies = [
    "alembic==1.13.3",
    "amqp==5.2.0",
    "annotated-types==0.7.0",
    "anyio==4.6.0",
    "billiard==4.2.1",
    "celery==5.4.0",
    "certifi==2024.8.30",
    "charset-normalizer==3.3.2",
    "click==8.1.7",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1",
    "click-repl==0.3.0",
    "eventkit==1.0.3",
    "exchange_calendars==4.5.6",
    "fastapi==0.115.0",
    "flower==2.0.1",
    "greenlet==3.1.1",
    "h11==0.14.0",
    "httptools==0.6.1",
    "humanize==4.10.0",
    "ib-insync==0.9.86",
    "idna==3.10",
    "kombu==5.4.2",
    "korean-lunar-calendar==0.3.1",
    "Mako==1.3.5",
    "MarkupSafe==2.1.5",
    "nest-asyncio==1.6.0",
    "numpy==2.1.1",
    "pandas==2.2.3",
    "pandas_market_calendars==4.4.1",
    "prometheus_client==0.21.0",
    "prompt_toolkit==3.0.47",
    "psutil==6.0.0",
    "psycopg2==2.9.9",
    "pydantic==2.9.2",
    "pydantic_core==2.23.4",
    "pyluach==2.2.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "pytz==2024.2",
    "PyYAML==6.0.2",
    "redis==5.0.8",
    "requests==2.32.3",
    "six==1.16.0",
    "sniffio==1.3.1",
    "SQLAlchemy==2.0.35",
    "starlette==0.38.6",
    "toolz==0.12.1",
    "tornado==6.4.1",
    "typing_extensions==4.12.2",
    "tzdata==2024.1",
    "urllib3==2.2.3",
    "uvicorn==0.30.6",
    "uvloop==0.20.0",
    "vine==5.1.0",
    "watchfiles==0.24.0",
    "wcwidth==0.2.13",
    "websockets==13.1",
    "pynacl",
    "pandas",
    "polars",
    "xlsxwriter",
    "matplotlib",
    "jupyter",
    "jupyter_bokeh",
    "coverage",
    "seaborn",
    "scikit-learn",
    "statsmodels",
    "tabulate",
    "panel",
    "watchfiles",
    "hvplot",
    "tqdm",
    "pykalman",
    "plotly",
    "yfinance",
    "xgboost",
]

[project.optional-dependencies]
test = ["pytest"]